fix: handle report not found
