+++ src/components/ReportGeneratedPoolDetail/index.tsx
@@ -1,5 +1,7 @@
 import React, { useMemo, createContext } from "react";
-import { useParams } from "react-router-dom";
+import { useHistory, useParams } from "react-router-dom";
+import { Box, styled } from "@mui/material";
+import { HiArrowLongLeft } from "react-icons/hi2";
 
 import { DeredistrationIcon, OperatorRewardIcon, PoolUpdateIcon, RegistrationIcon } from "src/commons/resources";
 import useFetch from "src/commons/hooks/useFetch";
@@ -16,6 +18,7 @@ import { SkeletonUI } from "../TokenDetail/TokenAnalytics/styles";
 import { getPoolEventList } from "../PoolLifecycle";
 import { Headline } from "../TabularView/StakeTab/styles";
 import CustomTooltip from "../commons/CustomTooltip";
+import NoRecord from "../commons/NoRecord";
 
 interface ITab {
   icon: React.FC;
@@ -60,6 +63,7 @@ const poolTabs: ITab[] = [
 
 const ReportGeneratedPoolDetailTabs = () => {
   const { reportId } = useParams<{ reportId: string }>();
+  const history = useHistory();
   const reportDetail = useFetch<IPoolReportList>(API.REPORT.POOL_REPORTED_DETAIL(reportId));
 
   const events = useMemo(() => {
@@ -87,32 +91,70 @@ const ReportGeneratedPoolDetailTabs = () => {
 
   const initTab = useMemo(() => (displayedTabs.length ? displayedTabs[0].key : undefined), [displayedTabs]);
 
+  if ((reportDetail.initialized && !reportDetail.data) || reportDetail.error) {
+    return <NoRecord />;
+  }
+
   return (
-    <ReportGeneratedPoolDetailContext.Provider
-      value={{
-        reportName: reportDetail.data?.reportHistory.reportName ?? "",
-        poolId: reportDetail.data?.poolView ?? ""
-      }}
-    >
-      {reportDetail.loading ? (
-        <SkeletonUI variant="rectangular" style={{ height: "400px" }} />
-      ) : (
-        <>
-          <CustomTooltip title={`${reportDetail.data?.reportHistory.reportName}`.replaceAll("-", " ")}>
-            <Headline data-testid="pool-report-name" collapsed={1}>
-              {`${reportDetail.data?.reportHistory.reportName}`.replaceAll("-", " ")}{" "}
-            </Headline>
-          </CustomTooltip>
-          <CustomTooltip title={`${reportDetail.data?.reportHistory.reportName}`.replaceAll("-", " ")}>
-            <Headline data-testid="pool-report-name">
-              {`${reportDetail.data?.reportHistory.reportName}`.replaceAll("-", " ")}{" "}
-            </Headline>
-          </CustomTooltip>
-          <StakeTab tabs={displayedTabs} initTab={initTab} />
-        </>
-      )}
-    </ReportGeneratedPoolDetailContext.Provider>
+    <Box>
+      <TopHeader>
+        <BackButton onClick={history.goBack}>
+          <HiArrowLongLeft />
+          <BackText>Back</BackText>
+        </BackButton>
+      </TopHeader>
+      <ReportGeneratedPoolDetailContext.Provider
+        value={{
+          reportName: reportDetail.data?.reportHistory.reportName ?? "",
+          poolId: reportDetail.data?.poolView ?? ""
+        }}
+      >
+        {reportDetail.loading ? (
+          <SkeletonUI variant="rectangular" style={{ height: "400px" }} />
+        ) : (
+          <>
+            <CustomTooltip title={`${reportDetail.data?.reportHistory.reportName}`.replaceAll("-", " ")}>
+              <Headline data-testid="pool-report-name" collapsed={1}>
+                {`${reportDetail.data?.reportHistory.reportName}`.replaceAll("-", " ")}{" "}
+              </Headline>
+            </CustomTooltip>
+            <CustomTooltip title={`${reportDetail.data?.reportHistory.reportName}`.replaceAll("-", " ")}>
+              <Headline data-testid="pool-report-name">
+                {`${reportDetail.data?.reportHistory.reportName}`.replaceAll("-", " ")}{" "}
+              </Headline>
+            </CustomTooltip>
+            <StakeTab tabs={displayedTabs} initTab={initTab} />
+          </>
+        )}
+      </ReportGeneratedPoolDetailContext.Provider>
+    </Box>
   );
 };
 
 export default ReportGeneratedPoolDetailTabs;
+
+const TopHeader = styled(Box)`
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+`;
+
+const BackButton = styled(Box)`
+  display: inline-flex;
+  align-items: center;
+  gap: 10px;
+  cursor: pointer;
+  ${({ theme }) => theme.breakpoints.down("md")} {
+    margin-top: 30px;
+    position: relative;
+    top: 5px;
+  }
+  ${({ theme }) => theme.breakpoints.down("sm")} {
+    margin-top: 0px;
+  }
+`;
+
+const BackText = styled("small")`
+  color: ${(props) => props.theme.palette.secondary.light};
+  font-weight: var(--font-weight-bold);
+`;
