+++ src/components/ReportGeneratedStakingDetail/index.tsx
@@ -1,5 +1,7 @@
 import React, { useMemo, createContext } from "react";
-import { useParams } from "react-router-dom";
+import { useHistory, useParams } from "react-router-dom";
+import { Box, styled } from "@mui/material";
+import { HiArrowLongLeft } from "react-icons/hi2";
 
 import {
   DelegationIcon,
@@ -23,6 +25,7 @@ import { getEventList } from "../StakekeySummary";
 import { SkeletonUI } from "../TokenDetail/TokenAnalytics/styles";
 import { Headline } from "../TabularView/StakeTab/styles";
 import CustomTooltip from "../commons/CustomTooltip";
+import NoRecord from "../commons/NoRecord";
 
 interface ITab {
   icon: React.FC;
@@ -74,6 +77,7 @@ const stackeTabs: ITab[] = [
 
 const ReportGeneratedStakingDetailTabs = () => {
   const { reportId } = useParams<{ reportId: string }>();
+  const history = useHistory();
   const reportDetail = useFetch<IReportStaking>(API.REPORT.STAKING_REPORTED_DETAIL(reportId));
 
   const events = useMemo(() => {
@@ -101,29 +105,67 @@ const ReportGeneratedStakingDetailTabs = () => {
   }, [events, reportDetail]);
   const initTab = useMemo(() => (displayedTabs.length ? displayedTabs[0].key : undefined), [displayedTabs]);
 
+  if ((reportDetail.initialized && !reportDetail.data) || reportDetail.error) {
+    return <NoRecord />;
+  }
+
   return (
-    <StakingDetailContext.Provider
-      value={{ stakeKey: reportDetail.data?.stakeKey ?? "", reportName: reportDetail.data?.reportName ?? "" }}
-    >
-      {reportDetail.loading ? (
-        <SkeletonUI variant="rectangular" style={{ height: "400px" }} />
-      ) : (
-        <>
-          <CustomTooltip title={`${reportDetail.data?.reportName}`.replaceAll("-", " ")}>
-            <Headline data-tesid="staking-report-name" collapsed={1}>
-              {`${reportDetail.data?.reportName}`.replaceAll("-", " ")}
-            </Headline>
-          </CustomTooltip>
-          <CustomTooltip title={`${reportDetail.data?.reportName}`.replaceAll("-", " ")}>
-            <Headline data-tesid="staking-report-name">
-              {`${reportDetail.data?.reportName}`.replaceAll("-", " ")}{" "}
-            </Headline>
-          </CustomTooltip>
-          <StakeTab tabs={displayedTabs} initTab={initTab} />
-        </>
-      )}
-    </StakingDetailContext.Provider>
+    <Box>
+      <TopHeader>
+        <BackButton onClick={history.goBack}>
+          <HiArrowLongLeft />
+          <BackText>Back</BackText>
+        </BackButton>
+      </TopHeader>
+      <StakingDetailContext.Provider
+        value={{ stakeKey: reportDetail.data?.stakeKey ?? "", reportName: reportDetail.data?.reportName ?? "" }}
+      >
+        {reportDetail.loading ? (
+          <SkeletonUI variant="rectangular" style={{ height: "400px" }} />
+        ) : (
+          <>
+            <CustomTooltip title={`${reportDetail.data?.reportName}`.replaceAll("-", " ")}>
+              <Headline data-tesid="staking-report-name" collapsed={1}>
+                {`${reportDetail.data?.reportName}`.replaceAll("-", " ")}
+              </Headline>
+            </CustomTooltip>
+            <CustomTooltip title={`${reportDetail.data?.reportName}`.replaceAll("-", " ")}>
+              <Headline data-tesid="staking-report-name">
+                {`${reportDetail.data?.reportName}`.replaceAll("-", " ")}{" "}
+              </Headline>
+            </CustomTooltip>
+            <StakeTab tabs={displayedTabs} initTab={initTab} />
+          </>
+        )}
+      </StakingDetailContext.Provider>
+    </Box>
   );
 };
 
 export default ReportGeneratedStakingDetailTabs;
+
+const TopHeader = styled(Box)`
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+`;
+
+const BackButton = styled(Box)`
+  display: inline-flex;
+  align-items: center;
+  gap: 10px;
+  cursor: pointer;
+  ${({ theme }) => theme.breakpoints.down("md")} {
+    margin-top: 30px;
+    position: relative;
+    top: 5px;
+  }
+  ${({ theme }) => theme.breakpoints.down("sm")} {
+    margin-top: 0px;
+  }
+`;
+
+const BackText = styled("small")`
+  color: ${(props) => props.theme.palette.secondary.light};
+  font-weight: var(--font-weight-bold);
+`;
