+++ src/pages/ReportGeneratedStakingDetail/index.tsx
@@ -1,27 +1,15 @@
 import { useEffect } from "react";
-import { useHistory } from "react-router-dom";
-import { HiArrowLongLeft } from "react-icons/hi2";
 import { Container } from "@mui/material";
 
 import ReportGeneratedDetailTabs from "src/components/ReportGeneratedStakingDetail";
 
-import { BackButton, BackText, TopHeader } from "./styles";
-
 const ReportGeneratedStakingDetail = () => {
-  const history = useHistory();
-
   useEffect(() => {
     document.title = "Stake Address Report Detail | Cardano Blockchain Explorer";
   }, []);
 
   return (
     <Container>
-      <TopHeader>
-        <BackButton onClick={history.goBack}>
-          <HiArrowLongLeft />
-          <BackText>Back</BackText>
-        </BackButton>
-      </TopHeader>
       <ReportGeneratedDetailTabs />
     </Container>
   );
